// src/renderer/src/components/inventarioYlogistica/historiales/HistorialDespachoDescarte/types/types.d.ts
/* eslint-disable prettier/prettier */

export type formType = {
    cliente:string
    nombreConductor: string
    telefono: string
    cedula:string
    remision:string
    tipoFruta:string
    "descarteLavado.descarteGeneral": string
    "descarteLavado.pareja": string
    "descarteLavado.balin": string
    "descarteEncerado.descarteGeneral": string
    "descarteEncerado.pareja": string
    "descarteEncerado.balin": string
    "descarteEncerado.extra": string
    "descarteEncerado.suelo": string
    "descarteEncerado.frutaNacional": string
}